# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory System API",
        "description": "Documentación de la API del sistema de inventarios",
        "contact": {
            "email": "dev@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "description": "Listar todos los productos ordenados descendente por Total\nTotal = Suma de cantidades en inventarios",
                "operationId": "70a301af8ea04ce2f2596a58758632cd",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto con inventario inicial",
                "description": "Crear un producto y asignar cantidad inicial en inventario",
                "operationId": "4ff8851cb6bfe92aa226cdc804917ff0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "maxLength": 300
                                    },
                                    "estado": {
                                        "type": "boolean"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    },
                                    "id_bodega": {
                                        "type": "integer"
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "producto": {
                                                    "$ref": "#/components/schemas/Producto"
                                                },
                                                "inventario_inicial": {
                                                    "$ref": "#/components/schemas/Inventario"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validación fallida"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Inventario": {
                "title": "Inventario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_bodega": {
                        "type": "integer"
                    },
                    "id_producto": {
                        "type": "integer"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_by": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Producto": {
                "title": "Producto",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string",
                        "nullable": true
                    },
                    "estado": {
                        "type": "boolean"
                    },
                    "created_by": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}